fixed an exception that triggers when resetting config.

+++ Config.cs
@@ -47,8 +47,8 @@ namespace p5rpc.events.saepov.Configuration
 
         [DisplayName("POV Model Major Id")]
         [Description("Allows the user to change which model POV will trigger for. THIS MAY NOT WORK CONSISTENLY FOR ALL MODELS.")]
-        [DefaultValue(1005)]
-        public short PovModelMajorId { get; set; } = 1005;
+        [DefaultValue((short)1005)]
+        public short PovModelMajorId { get; set; } = (short)1005;
     }
 
     /// <summary>
